+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog].
 
 ## Unreleased
+### Enhancements
+* Improved evil-mode's search history integration so it will now use
+  the translations defined in `ctrlf-style-alist` ([#124]).
+
+[#124]: https://github.com/radian-software/ctrlf/pull/124
+
 ### Bugs fixed
 * Previously `C-g` would fail to reset point to its original position
   in some cases. This has been fixed ([#110], [#114]).